+++ domain/src/main/java/dev/yjyoon/kwnotice/domain/usecase/favorite/AddFavoriteUseCase.kt
@@ -1,4 +1,4 @@
-package dev.yjyoon.kwnotice.domain.usecase
+package dev.yjyoon.kwnotice.domain.usecase.favorite
 
 import dev.yjyoon.kwnotice.domain.model.Favorite
 import dev.yjyoon.kwnotice.domain.repository.FavoriteRepository
