+++ domain/src/main/java/dev/yjyoon/kwnotice/domain/usecase/favorite/GetFavoriteSwIdListUseCase.kt
@@ -1,12 +1,12 @@
-package dev.yjyoon.kwnotice.domain.usecase
+package dev.yjyoon.kwnotice.domain.usecase.favorite
 
 import dev.yjyoon.kwnotice.domain.repository.FavoriteRepository
 import javax.inject.Inject
 
-class GetFavoriteKwIdListUseCase @Inject constructor(
+class GetFavoriteSwIdListUseCase @Inject constructor(
     private val favoriteRepository: FavoriteRepository
 ) {
 
     suspend operator fun invoke(): Result<List<Long>> =
-        favoriteRepository.getFavoriteKwHomeIds()
+        favoriteRepository.getFavoriteSwCentralIds()
 }
