+++ domain/src/main/java/dev/yjyoon/kwnotice/domain/usecase/settings/UnsubscribeFcmTopicUseCase.kt
@@ -1,13 +1,13 @@
-package dev.yjyoon.kwnotice.domain.usecase
+package dev.yjyoon.kwnotice.domain.usecase.settings
 
 import dev.yjyoon.kwnotice.domain.model.FcmTopic
 import dev.yjyoon.kwnotice.domain.repository.SettingsRepository
 import javax.inject.Inject
 
-class SubscribeFcmTopicUseCase @Inject constructor(
+class UnsubscribeFcmTopicUseCase @Inject constructor(
     private val settingsRepository: SettingsRepository
 ) {
 
     suspend operator fun invoke(topic: FcmTopic): Result<Unit> =
-        settingsRepository.subscribeFcmTopic(topic)
+        settingsRepository.unsubscribeFcmTopic(topic)
 }
