+++ presentation/src/main/java/dev/yjyoon/kwnotice/presentation/ui/notice/NoticeViewModel.kt
@@ -4,12 +4,12 @@ import android.util.Log
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.yjyoon.kwnotice.domain.model.Notice
 import dev.yjyoon.kwnotice.domain.model.toFavorite
-import dev.yjyoon.kwnotice.domain.usecase.AddFavoriteUseCase
-import dev.yjyoon.kwnotice.domain.usecase.DeleteFavoriteUseCase
-import dev.yjyoon.kwnotice.domain.usecase.GetFavoriteKwIdListUseCase
-import dev.yjyoon.kwnotice.domain.usecase.GetFavoriteSwIdListUseCase
-import dev.yjyoon.kwnotice.domain.usecase.GetKwHomeNoticeListUseCase
-import dev.yjyoon.kwnotice.domain.usecase.GetSwCentralNoticeListUseCase
+import dev.yjyoon.kwnotice.domain.usecase.favorite.AddFavoriteUseCase
+import dev.yjyoon.kwnotice.domain.usecase.favorite.DeleteFavoriteUseCase
+import dev.yjyoon.kwnotice.domain.usecase.favorite.GetFavoriteKwIdListUseCase
+import dev.yjyoon.kwnotice.domain.usecase.favorite.GetFavoriteSwIdListUseCase
+import dev.yjyoon.kwnotice.domain.usecase.notice.GetKwHomeNoticeListUseCase
+import dev.yjyoon.kwnotice.domain.usecase.notice.GetSwCentralNoticeListUseCase
 import dev.yjyoon.kwnotice.presentation.ui.base.BaseViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
