+++ presentation/src/main/java/dev/yjyoon/kwnotice/presentation/ui/splash/SplashViewModel.kt
@@ -2,8 +2,8 @@ package dev.yjyoon.kwnotice.presentation.ui.splash
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.yjyoon.kwnotice.domain.model.FcmTopic
-import dev.yjyoon.kwnotice.domain.usecase.CheckFirstLaunchUseCase
-import dev.yjyoon.kwnotice.domain.usecase.SubscribeFcmTopicUseCase
+import dev.yjyoon.kwnotice.domain.usecase.settings.CheckFirstLaunchUseCase
+import dev.yjyoon.kwnotice.domain.usecase.settings.SubscribeFcmTopicUseCase
 import dev.yjyoon.kwnotice.presentation.fcm.FcmSubscription
 import dev.yjyoon.kwnotice.presentation.ui.base.BaseViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
